oarn/requirements.txt
Django==1.7
defusedxml==0.4.1
django-filter==0.10.0
djangorestframework==3.1.2
djangorestframework-xml==1.0.1
djoser==0.3.0
fuzzywuzzy==0.5.0
mccabe==0.3
mysqlclient==1.3.6
pep8==1.5.7
pyflakes==0.8.1
python-Levenshtein==0.12.0
selenium==2.45.0
xlrd==0.9.3
